Refactor split yaep.c
